fix(#4120): brand reified builtin constructor carriers callable + constructible
A reified builtin constructor (Set, Array, TypeError, RegExp, ...) is backed
in standalone by a plain $Object singleton (#3006/#2907). The typeof natives
classify by ref.test over closure wrapper structs, so that carrier answered
"object" - `typeof f === "function"` was FALSE for every builtin that had
reached a value slot, and test262's isConstructor(f) threw
"isConstructor invoked with a non-function value" before testing anything.
typeof cannot throw, so this was a SILENT wrong answer.

Measured standalone via a one-parameter indirection (the in-place spelling is
constant-folded and never touches the carrier):
  typeof Set / Map / Array / Object / TypeError / RegExp / Weak* -> "object"
  typeof Math / JSON / Reflect -> "object"  (correct, untouched)
  typeof Array.from / Object.keys / Math.max -> "function" (already correct)

Fix, at the carrier-representation seam:
  - OBJ_FLAG_CALLABLE (0x10) / OBJ_FLAG_CONSTRUCTOR (0x20) internal-slot bits
    on the carrier's existing $Object.flags word. Representation, identity and
    own properties are unchanged; every other flags reader masks only its own
    bit. Two bits, not one: isNaN is callable and NOT constructible, so a
    later callable-only brand must not be able to claim [[Construct]].
  - one shared predicate arm in __typeof_function / __typeof_object /
    __typeof, and in __reflect_is_constructor (§7.2.4 IsConstructor).
  - the brand is applied only where a spec ctor arity exists, so the
    Math/JSON/Reflect namespace carriers stay "object" as the spec requires.

fillStandaloneTypeofClosureArms moves verbatim out of closure-exports.ts into
its own subsystem module (typeof-natives-finalize.ts) rather than growing a
god file; closure-exports.ts re-exports it so index.ts is untouched.


✓
